Add files via upload
This was a project done my advance econometrics on a data replication of Hughes 2006 research on the short-run price elasticity of gasoline demand. This document show a replication the tables and graph done in R on the data but instead it focuses on the data collected from 1975 to 2021. I currently working on a research project which will be focusing on this data and table looking at what the short-run elasticity is.